Fix: Having different files for each svg
