+++ frontend/src/assets/icons/CopyIcon.jsx
@@ -1,0 +1,31 @@
+import { defaultSvgProps } from './constants';
+
+export function CopyIcon({ size = 20, ...props }) {
+  return (
+    <svg
+      {...defaultSvgProps}
+      viewBox="0 0 24 24"
+      width={size}
+      height={size}
+      {...props}
+    >
+      <rect
+        x="9"
+        y="9"
+        width="13"
+        height="13"
+        rx="2"
+        ry="2"
+        stroke="currentColor"
+        strokeWidth="2"
+      />
+      <path
+        d="M5 15H4a2 2 0 0 1-2-2V4a2 2 0 0 1 2-2h9a2 2 0 0 1 2 2v1"
+        stroke="currentColor"
+        strokeWidth="2"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+      />
+    </svg>
+  );
+}
