+++ frontend/src/assets/icons/RocketIcon.jsx
@@ -1,0 +1,23 @@
+import { defaultSvgProps } from './constants';
+
+export function RocketIcon({ size = 20, ...props }) {
+  return (
+    <svg
+      {...defaultSvgProps}
+      viewBox="0 0 24 24"
+      width={size}
+      height={size}
+      {...props}
+    >
+      <circle cx="12" cy="12" r="10" stroke="currentColor" strokeWidth="2" fill="none" />
+      <path
+        d="M9 12l2 2 4-4"
+        stroke="currentColor"
+        strokeWidth="2"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+        fill="none"
+      />
+    </svg>
+  );
+}
