+++ frontend/src/assets/icons/UnfoldMoreIcon.jsx
@@ -1,0 +1,22 @@
+import { defaultSvgProps } from './constants';
+
+export function UnfoldMoreIcon({ size = 20, ...props }) {
+  return (
+    <svg
+      {...defaultSvgProps}
+      viewBox="0 0 24 24"
+      width={size}
+      height={size}
+      {...props}
+    >
+      <path
+        d="M6 9l6 6 6-6"
+        stroke="currentColor"
+        strokeWidth="2"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+        fill="none"
+      />
+    </svg>
+  );
+}
