+++ frontend/src/assets/icons/constants.js
@@ -1,0 +1,9 @@
+/**
+ * Shared SVG props for all icon components.
+ * These props ensure consistent accessibility and styling across icons.
+ */
+export const defaultSvgProps = {
+  "aria-hidden": true,
+  focusable: "false",
+  fill: "none",
+};
